BANP: Move NamedPorts into extended test suite
