+++ conformance/tests/baseline-admin-network-policy-core-ingress-tcp-rules.go
@@ -33,7 +33,6 @@ import (
 func init() {
 	ConformanceTests = append(ConformanceTests,
 		BaselineAdminNetworkPolicyIngressTCP,
-		BaselineAdminNetworkPolicyIngressNamedPort,
 	)
 }
 
@@ -147,58 +146,5 @@ var BaselineAdminNetworkPolicyIngressTCP = suite.ConformanceTest{
 				serverPod.Status.PodIP, int32(8080), s.TimeoutConfig.RequestTimeout, true)
 			assert.True(t, success)
 		})
-	},
-}
-
-var BaselineAdminNetworkPolicyIngressNamedPort = suite.ConformanceTest{
-	ShortName:   "BaselineAdminNetworkPolicyIngressNamedPort",
-	Description: "Tests support for ingress traffic on a named port using baseline admin network policy API based on a server and client model",
-	Features: []suite.SupportedFeature{
-		suite.SupportBaselineAdminNetworkPolicy,
-		suite.SupportBaselineAdminNetworkPolicyNamedPorts,
-	},
-	Manifests: []string{"base/baseline_admin_network_policy/core-ingress-tcp-rules.yaml"},
-	Test: func(t *testing.T, s *suite.ConformanceTestSuite) {
-
-		t.Run("Should support an 'allow-ingress' policy for named port", func(t *testing.T) {
-			ctx, cancel := context.WithTimeout(context.Background(), s.TimeoutConfig.GetTimeout)
-			defer cancel()
-			// This test uses `default` BANP
-			// harry-potter-1 is our server pod in gryffindor namespace
-			serverPod := &v1.Pod{}
-			err := s.Client.Get(ctx, client.ObjectKey{
-				Namespace: "network-policy-conformance-gryffindor",
-				Name:      "harry-potter-1",
-			}, serverPod)
-			require.NoErrorf(t, err, "unable to fetch the server pod")
-			banp := &v1alpha1.BaselineAdminNetworkPolicy{}
-			err = s.Client.Get(ctx, client.ObjectKey{
-				Name: "default",
-			}, banp)
-			require.NoErrorf(t, err, "unable to fetch the baseline admin network policy")
-			namedPortRule := banp.DeepCopy().Spec.Ingress[3]
-			webPort := "web"
-			// rewrite the tcp port 80 rule as named port rule
-			namedPortRule.Ports = &[]v1alpha1.AdminNetworkPolicyPort{
-				{
-					NamedPort: &webPort,
-				},
-			}
-			banp.Spec.Ingress[3] = namedPortRule
-			err = s.Client.Update(ctx, banp)
-			require.NoErrorf(t, err, "unable to update the baseline admin network policy")
-			// cedric-diggory-0 is our client pod in hufflepuff namespace
-			// ensure ingress is ALLOWED from hufflepuff to gryffindor at at the web port, which is defined as TCP at port 80 in pod spec
-			// ingressRule at index3 should take effect
-			success := kubernetes.PokeServer(t, s.ClientSet, &s.KubeConfig, "network-policy-conformance-hufflepuff", "cedric-diggory-0", "tcp",
-				serverPod.Status.PodIP, int32(80), s.TimeoutConfig.RequestTimeout, true)
-			assert.True(t, success)
-			// cedric-diggory-1 is our client pod in hufflepuff namespace
-			// ensure ingress is DENIED from hufflepuff to gryffindor for rest of the traffic; ingressRule at index4 should take effect
-			success = kubernetes.PokeServer(t, s.ClientSet, &s.KubeConfig, "network-policy-conformance-hufflepuff", "cedric-diggory-1", "tcp",
-				serverPod.Status.PodIP, int32(8080), s.TimeoutConfig.RequestTimeout, false)
-			assert.True(t, success)
-		})
-
 	},
 }
